==> Notebooks/04_Prepare_Data_For_Modeling.ipynb <==
# 04 Prepare Data For Modeling

### Purpose of Notebook
- Split features from text data
- Scale feature data
- Tokenize text data
- Combine feature and text data
- Export X data for use later in workflow

## Imports

import numpy as np
import pandas as pd
import pickle
from sklearn.feature_extraction.text import TfidfVectorizer, CountVectorizer
from sklearn.feature_extraction.text import ENGLISH_STOP_WORDS
from sklearn.preprocessing import StandardScaler
from string import punctuation

## Pull in X_train and X_test data

with open('../Data/X_train.pkl', 'rb') as f:
    X_train = pickle.load(f)

with open('../Data/X_test.pkl', 'rb') as f:
    X_test = pickle.load(f)

## Create X DataFrame without Text Data 

X_train_features = X_train.drop(labels=['selftext','title'], axis=1).copy()
X_test_features = X_test.drop(labels=['selftext','title'], axis=1).copy()

## Fit Standard Scalar and Transform Feature Data

ss = StandardScaler()
ss.fit(X_train_features);

X_train_features_sc = pd.DataFrame(ss.transform(X_train_features), columns=X_train_features.columns)
X_test_features_sc = pd.DataFrame(ss.transform(X_test_features), columns=X_test_features.columns)

## Save Standard Scaler for later use

with open('../Objects/ss.pkl', 'wb') as f:
    pickle.dump(ss, f)

## Extract Text Data

X_train_text = list(X_train.title + ' ' + X_train.selftext)
X_test_text = list(X_test.title + ' ' + X_test.selftext)

## Setup additional stopwords

_stopwords = set(list(ENGLISH_STOP_WORDS)+list(punctuation))

## Fit Vectorizer and Transform Text Data

tvec = TfidfVectorizer(stop_words=_stopwords, min_df=2, max_df=0.5, ngram_range=(1,2))
tvec.fit(X_train_text);

X_train_text_sp = pd.SparseDataFrame(tvec.transform(X_train_text), 
                                      columns = tvec.get_feature_names(), default_fill_value=0)

X_test_text_sp = pd.SparseDataFrame(tvec.transform(X_test_text), 
                                      columns = tvec.get_feature_names(), default_fill_value=0)

## Save Vectorizer for later use

with open('../Objects/tvec.pkl','wb') as f:
    pickle.dump(tvec, f)

## Combine Features with Text Data
### Try to leave out additional features to see if it will improve the model

feature_cols = X_train_features_sc.columns

X_train_text_sp[feature_cols] = X_train_features_sc[feature_cols]
X_test_text_sp[feature_cols] = X_test_features_sc[feature_cols]

## Export Processed Feature Data

with open('../Data/X_train_clean.pkl', 'wb') as f:
    pickle.dump(X_train_text_sp, f)
    
with open('../Data/X_test_clean.pkl', 'wb') as f:
    pickle.dump(X_test_text_sp, f)
