==> Notebooks/05_Create_Baseline_Models.ipynb <==
# 05 Create Baseline Models

### Purpose of Notebook
- Build a bunch of quick-to-train models:
    - Logistic Regression
    - K Nearest Neighbors
    - Random Forest
    - AdaBoost with GridSearch for CV
    
    
- Put results of each model into a table and review

## Imports

import pandas as pd
import numpy as np
import pickle
from sklearn.model_selection import train_test_split, GridSearchCV
from sklearn.linear_model import LogisticRegressionCV, LogisticRegression
from sklearn.neighbors import KNeighborsClassifier
from sklearn.ensemble import RandomForestClassifier
import random
from matplotlib import pyplot as plt
from sklearn.metrics import confusion_matrix
from sklearn.ensemble import AdaBoostClassifier, GradientBoostingClassifier, VotingClassifier
from sklearn.svm import SVC

random.seed(42)

def get_metrics(y_true, y_predict, print_scores = True):
    matrix_def = [['tn','fp'], ['fn','tp']]
    matrix = confusion_matrix(y_true, y_predict)
    tn, fp, fn, tp = matrix.ravel()
    accuracy = (tp+tn)/(tn+fp+fn+tp)
    misclass = 1-accuracy
    sensitivity = tp/(tp+fn)
    specificity = tn/(tn+fp)
    precision = tp/(tp+fp)
    if print_scores:
        print('Matrix Definition')
        print(np.array(matrix_def))
        print('')
        print('Confusion Matrix')
        print(matrix)
        print('')
        print('METRICS')
        print(f'accuracy: {accuracy}')
        print(f'misclass: {misclass}')
        print(f'sensitivity: {sensitivity}')
        print(f'specificity: {specificity}')
    else:
        return accuracy, misclass, sensitivity, specificity, precision

## Pull in train and test data

with open('../Data/X_train_clean.pkl', 'rb') as f:
    X_train = pickle.load(f)
    
with open('../Data/X_test_clean.pkl', 'rb') as f:
    X_test = pickle.load(f)
    
with open('../Data/y_train.pkl', 'rb') as f:
    y_train = pickle.load(f)
    
with open('../Data/y_test.pkl', 'rb') as f:
    y_test = pickle.load(f)

## Create DataFrame to hold results

results = pd.DataFrame(columns=['model','train_score', 'test_score'])

## Logistic Regression

lr = LogisticRegression()
lr.fit(X_train, y_train)
print('Train:', lr.score(X_train,y_train))
print('Test:', lr.score(X_test,y_test))

get_metrics(y_test, lr.predict(X_test))

counter = len(results) + 1
results.loc[counter, 'model'] = 'Logistic Regression'
results.loc[counter, 'train_score'] = lr.score(X_train,y_train)
results.loc[counter, 'test_score'] = lr.score(X_test,y_test)

coefficients = pd.DataFrame(lr.coef_[0], 
                            index=X_train.columns,
                            columns=['coefficients']).sort_values('coefficients', ascending=False)

coefficients.head(10)

coefficients.sort_values('coefficients').head(10)

## K Nearest Neighbors

knn = KNeighborsClassifier()
knn.fit(X_train, y_train)
print('Train:', knn.score(X_train,y_train))
print('Test:', knn.score(X_test,y_test))

get_metrics(y_test, knn.predict(X_test))

counter = len(results) + 1
results.loc[counter, 'model'] = 'K Nearest Neighbors'
results.loc[counter, 'train_score'] = knn.score(X_train,y_train)
results.loc[counter, 'test_score'] = knn.score(X_test,y_test)

## Random Forest

rf = RandomForestClassifier()
rf.fit(X_train, y_train)
print('Train:', rf.score(X_train,y_train))
print('Test:',rf.score(X_test,y_test))

get_metrics(y_test, rf.predict(X_test))

counter = len(results) + 1
results.loc[counter, 'model'] = 'Random Forest'
results.loc[counter, 'train_score'] = rf.score(X_train,y_train)
results.loc[counter, 'test_score'] = rf.score(X_test,y_test)

## AdaBoost with GridSearch for CV

ada = AdaBoostClassifier(random_state=42)
ada_params = {}
gs = GridSearchCV(ada, param_grid=ada_params)
gs.fit(X_train, y_train)
print('Best Score:', gs.best_score_)
print('Best Parameters:', gs.best_params_)
print('Test:',gs.score(X_test,y_test))

feature_importance = pd.DataFrame(gs.best_estimator_.feature_importances_, 
                                  index=X_train.columns,
                                  columns=['coefficients']).sort_values('coefficients', 
                                                                        ascending=False)

feature_importance.head(10)

get_metrics(y_test, gs.predict(X_test))

counter = len(results) + 1
results.loc[counter, 'model'] = 'AdaBoost'
results.loc[counter, 'train_score'] = gs.score(X_train,y_train)
results.loc[counter, 'test_score'] = gs.score(X_test,y_test)

## View Baseline Model Results

results
