==> Notebooks/05B_Fit_And_Save_Complex_Models.ipynb <==
# 05 Fit and Save Complex Models Separately

### Purpose of Notebook
- Try some of the more complex models:
- Export fit models for later review

## Imports

import pandas as pd
import numpy as np
import pickle
from sklearn.model_selection import train_test_split, GridSearchCV, RandomizedSearchCV
from sklearn.linear_model import LogisticRegressionCV, LogisticRegression
from sklearn.neighbors import KNeighborsClassifier
from sklearn.ensemble import RandomForestClassifier, BaggingClassifier
from sklearn.ensemble import GradientBoostingClassifier, VotingClassifier
from sklearn.tree import DecisionTreeClassifier
import random
from matplotlib import pyplot as plt
from sklearn.metrics import confusion_matrix
from sklearn.ensemble import AdaBoostClassifier, GradientBoostingClassifier, VotingClassifier
from sklearn.svm import SVC
import pickle

random.seed(42)

## Pull in train and test data

with open('../Data/X_train_clean.pkl', 'rb') as f:
    X_train = pickle.load(f)
    
with open('../Data/X_test_clean.pkl', 'rb') as f:
    X_test = pickle.load(f)
    
with open('../Data/y_train.pkl', 'rb') as f:
    y_train = pickle.load(f)
    
with open('../Data/y_test.pkl', 'rb') as f:
    y_test = pickle.load(f)

## Random Forest with GridSearch

%%time
rf = RandomForestClassifier(random_state=42, min_samples_split=2, min_samples_leaf=1)
rf_params = {
    'n_estimators': [10, 30, 50, 100],
    'min_samples_split': [2,3],
    'min_samples_leaf': [1,2]
}
rf_gs = GridSearchCV(rf, param_grid=rf_params, verbose=2, n_jobs=3)
rf_gs.fit(X_train, y_train)

with open('../Models/RandomForest_gs.pkl', 'wb') as f:
    pickle.dump(rf_gs, f)

## Try SVM

%%time
sv = SVC()
param_grid = {
    'gamma': np.logspace(-5, 2, 20)
}

gs_sv = GridSearchCV(sv, param_grid,scoring='accuracy', verbose=1, n_jobs=3)
gs_sv.fit(X_train, y_train)
print('Best Score:', gs_sv.best_score_)
print('Best Parameters:', gs_sv.best_params_)
print('Test:',gs_sv.score(X_test,y_test))

## Logistic Regression with Grid Search

%%time
lr = LogisticRegression(random_state=42)
lr_params = {
    'C': [0.1, 1, 10],
    'penalty': ['l2']
}
lr_gs = GridSearchCV(lr, param_grid=lr_params, verbose=2, n_jobs=3)
lr_gs.fit(X_train, y_train)

print('Train:', lr_gs.score(X_train,y_train))
print('Test:', lr_gs.score(X_test,y_test))

lr_gs.best_params_

with open('../Models/LogisticRegression_gs.pkl', 'wb') as f:
    pickle.dump(lr_gs, f)

## Bagging Classifier with Decision Trees

%%time
bag = BaggingClassifier(random_state=42, base_estimator=DecisionTreeClassifier())
bag_params = {
    'n_estimators':[10, 15],
    'base_estimator__max_depth': list(range(4,10,2)),
    'base_estimator__min_samples_split': [2,3]
}
bag_gs = GridSearchCV(bag, param_grid=bag_params, verbose=2, n_jobs=3)
bag_gs.fit(X_train, y_train)

print('Best Params:', bag_gs.best_params_)
print('Train:', bag_gs.score(X_train,y_train))
print('Test:', bag_gs.score(X_test,y_test))

with open('../Models/Bag_Trees_gs.pkl', 'wb') as f:
    pickle.dump(bag_gs, f)

## Gradient Boosting with Grid Search

%%time
gb = GradientBoostingClassifier(random_state=42)
gb_params = {
    'learning_rate': [0.05, 0.1, 0.3],
    'max_depth': [3,4,5]
}
gb_gs = GridSearchCV(gb, param_grid=gb_params, verbose=2, n_jobs=3 )
gb_gs.fit(X_train, y_train)

print('Best Params:', gb_gs.best_params_)
print('Train:', gb_gs.score(X_train,y_train))
print('Test:', gb_gs.score(X_test,y_test))

with open('../Models/GradientBoost_gs.pkl', 'wb') as f:
    pickle.dump(gb_gs, f)

## AdaBoost with Grid Search

%%time
ada = AdaBoostClassifier(random_state=42)
ada_params = {
    'n_estimators':[50,100],
    'learning_rate': [0.5, 1.0]
}
ada_gs = GridSearchCV(ada, param_grid=ada_params, verbose=2, n_jobs=3)
ada_gs.fit(X_train, y_train)


print('Best Score:', ada_gs.best_score_)
print('Best Parameters:', ada_gs.best_params_)
print('Train:',ada_gs.score(X_train,y_train))
print('Test:',ada_gs.score(X_test,y_test))

with open('../Models/AdaBoost_gs.pkl','wb') as f:
    pickle.dump(ada_gs, f)

## Final Try Before Moving On:  Voting Classifier with the best Models

%%time
vote = VotingClassifier([
    ('ada', AdaBoostClassifier(learning_rate=0.5)),
    ('rf', RandomForestClassifier(n_estimators=100)),
    ('lr', LogisticRegression(C=10)),
    ('gb', GradientBoostingClassifier(max_depth=5))
])
vote_params = {
    'weights': [[0.25,0.25,0.25, 0.25],
                [0.4, 0.4, 0.1, 0.1],
                [0.1, 0.1, 0.4, 0.4]]
}
vote_gs = GridSearchCV(vote, param_grid=vote_params, verbose=2, n_jobs=2)
vote_gs.fit(X_train, y_train)

print('Best Score:', vote_gs.best_score_)
print('Best Parameters:', vote_gs.best_params_)
print('Test:',vote_gs.score(X_train,y_train))
print('Test:',vote_gs.score(X_test,y_test))
