==> Notebooks/05C_Score_Complex_Models.ipynb <==
# 05C Score Complex Models

## Imports

import pickle
import pandas as pd

## Pull in Train and Test Data

with open('../Data/X_train_clean.pkl', 'rb') as f:
    X_train = pickle.load(f)
    
with open('../Data/X_test_clean.pkl', 'rb') as f:
    X_test = pickle.load(f)
    
with open('../Data/y_train.pkl', 'rb') as f:
    y_train = pickle.load(f)
    
with open('../Data/y_test.pkl', 'rb') as f:
    y_test = pickle.load(f)

## Create DataFrame to hold results

pd.set_option('display.max_colwidth', -1)

results = pd.DataFrame(columns=['model','train_score', 'test_score', 'best_params'])

## RandomForest_gs

with open('../Models/RandomForest_gs.pkl', 'rb') as f:
    rf_gs = pickle.load(f)

counter = len(results) + 1
results.loc[counter, 'model'] = 'RandomForest_gs'
results.loc[counter, 'train_score'] = rf_gs.score(X_train,y_train)
results.loc[counter, 'test_score'] = rf_gs.score(X_test,y_test)
results.loc[counter, 'best_params'] = str(rf_gs.best_params_)

## GradientBoost_gs

with open('../Models/GradientBoost_gs.pkl', 'rb') as f:
    gb_gs = pickle.load(f)

counter = len(results) + 1
results.loc[counter, 'model'] = 'GradientBoost_gs'
results.loc[counter, 'train_score'] = gb_gs.score(X_train,y_train)
results.loc[counter, 'test_score'] = gb_gs.score(X_test,y_test)
results.loc[counter, 'best_params'] = str(gb_gs.best_params_)

## AdaBoost_gs

with open('../Models/AdaBoost_gs.pkl', 'rb') as f:
    ada_gs = pickle.load(f)

counter = len(results) + 1
results.loc[counter, 'model'] = 'AdaBoost_gs'
results.loc[counter, 'train_score'] = ada_gs.score(X_train,y_train)
results.loc[counter, 'test_score'] = ada_gs.score(X_test,y_test)
results.loc[counter, 'best_params'] = str(ada_gs.best_params_)

## Bag_Trees_gs

with open('../Models/Bag_Trees_gs.pkl', 'rb') as f:
    bag_gs = pickle.load(f)

counter = len(results) + 1
results.loc[counter, 'model'] = 'BagTrees_gs'
results.loc[counter, 'train_score'] = bag_gs.score(X_train,y_train)
results.loc[counter, 'test_score'] = bag_gs.score(X_test,y_test)
results.loc[counter, 'best_params'] = str(bag_gs.best_params_)

coefficients = pd.DataFrame(lr.coef_[0], 
                            index=X_train.columns,
                            columns=['coefficients']).sort_values('coefficients', ascending=False)

## LogisticRegression_gs

with open('../Models/LogisticRegression_gs.pkl', 'rb') as f:
    lr_gs = pickle.load(f)

counter = len(results) + 1
results.loc[counter, 'model'] = 'LogisticRegression_gs'
results.loc[counter, 'train_score'] = lr_gs.score(X_train,y_train)
results.loc[counter, 'test_score'] = lr_gs.score(X_test,y_test)
results.loc[counter, 'best_params'] = str(lr_gs.best_params_)

results
